+++ src/index.js
@@ -4,6 +4,10 @@ const logger = new Logger('graceful-kill');
 
 export default (process, SIGTERMTimeout: number = 4000) =>
   new Promise(resolve => {
+    if (process.exitCode !== null || process.signalCode !== null) {
+      logger.warn('process already exited');
+      return resolve();
+    }
     const killTimeout = setTimeout(() => {
       logger.warn('timeout: sending SIGKILL...');
       process.kill('SIGKILL');
